+++ example/src/main/java/dev/enro/example/destinations/result/flow/managed/CreateSentenceManagedFlow.kt
@@ -7,8 +7,8 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
-import dev.enro.annotations.NavigationDestination
 import dev.enro.annotations.ExperimentalEnroApi
+import dev.enro.annotations.NavigationDestination
 import dev.enro.core.NavigationKey
 import dev.enro.core.closeWithResult
 import dev.enro.core.container.EmptyBehavior
@@ -48,10 +48,6 @@ class ManagedFlowViewModel(
             navigation.closeWithResult(it)
         }
     )
-
-    init {
-        flow.next()
-    }
 }
 
 @NavigationDestination(CreateSentenceManagedFlow::class)